Muevo tests de validación a test_validation.py
